app/gen/source/requirements.txt
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Jinja2==2.10
psycopg2-binary==2.8.4
SQLAlchemy==1.3.10
autopep8==1.4.4
Flask-WTF==0.14.2
Pillow-6.2.1